commit 7

+++ ControlBlockchain/BlockchainVersionControl.Tests/VersionControlBlockchainTests.cs
@@ -6,7 +6,7 @@ using System.IO;
 
 //Piotr Bacior 15 722 WSEI Kraków - Zadanie dodatkowe - Blockchain
 
-// Testy jednostkowe klasy VersionControlBlockchain - każdy test sprawdza inny aspekt działania blockchaina wersji dokumentu
+//Testy jednostkowe klasy VersionControlBlockchain - każdy test sprawdza inny aspekt działania blockchaina wersji dokumentu
 namespace BlockchainVersionControl.Tests
 {
     [TestClass]
@@ -22,18 +22,18 @@ namespace BlockchainVersionControl.Tests
             blockchain = new VersionControlBlockchain("Linia 1\nLinia 2");
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie, czy blockchain został poprawnie zainicjalizowany i czy zawiera dokładnie jeden blok genesis
+        //Definiujemy test inicjalizacji: sprawdzamy, czy utworzony został blok początkowy (genesis)
         [TestMethod]
-        public void CzyTworzyBlokGenesis()
+        public void CzyTworzyBlokPoczatkowy()
         {
-            //Act & Assert
+            //Assert
             Assert.AreEqual(1, blockchain.Chain.Count);
             Assert.AreEqual(0, blockchain.Chain[0].Index);
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie, czy dodanie nowej wersji dokumentu do blockchaina działa poprawnie
+        //Definiujemy teraz test dodania nowej wersji dokumentu
         [TestMethod]
-        public void CzyDodajeNowaWersje()
+        public void CzyDodajeWersje()
         {
             //Act
             blockchain.AddNewVersion("Linia 1\nLinia Zmieniona");
@@ -42,9 +42,9 @@ namespace BlockchainVersionControl.Tests
             Assert.AreEqual(2, blockchain.Chain.Count);
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy GetLatestVersionBlock zwraca ostatni blok w łańcuchu
+        //Definiujemy test zwracania ostatniego bloku w łańcuchu
         [TestMethod]
-        public void CzyZwracaNajnowszyBlok()
+        public void CzyZwracaOstatniBlok()
         {
             //Arrange
             blockchain.AddNewVersion("Nowy tekst");
@@ -56,9 +56,9 @@ namespace BlockchainVersionControl.Tests
             Assert.AreEqual(1, najnowszy.Index);
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy GetDocumentVersion zwraca poprawną treść dokumentu w danej wersji 
+        //Przechodzimu to zdefiniowania testu odtworzenia pełnej wersji dokumentu na podstawie wskazanej wersji
         [TestMethod]
-        public void CzyRekonstruujeWersjeDokumentu()
+        public void CzyRekonstruujeWersje()
         {
             //Arrange
             blockchain.AddNewVersion("Linia 1\nLinia Zmieniona");
@@ -70,9 +70,9 @@ namespace BlockchainVersionControl.Tests
             StringAssert.Contains(wynik, "Linia Zmieniona");
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy GetDiffBetweenVersions zwraca poprawne różnice między wersjami
+        //Definiujemy test generowania różnic pomiędzy dwiema wersjami
         [TestMethod]
-        public void CzyWyznaczaPoprawnyDiff()
+        public void CzyGenerujeDiffMiedzyWersjami()
         {
             //Arrange
             blockchain.AddNewVersion("Linia 1\nNowa linia");
@@ -80,24 +80,24 @@ namespace BlockchainVersionControl.Tests
             //Act
             var diff = blockchain.GetDiffBetweenVersions(0, 1);
 
-            //Assert.IsNotNull(diff);
+            //Assert
             StringAssert.Contains(diff, "+ Nowa linia");
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy IsChainValid zwraca true dla niezmodyfikowanego łańcucha
+        //Definujemy test poprawności łańcucha w przypadku braku manipulacji
         [TestMethod]
-        public void CzySprawdzaPoprawnoscLancuchaGdyPoprawny()
+        public void CzyLancuchJestPoprawny()
         {
-            //Arrange i act 
+            //Arrange
             blockchain.AddNewVersion("Linia 1\nNowa linia");
 
             //Assert
             Assert.IsTrue(blockchain.IsChainValid());
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy IsChainValid zwraca false dla zmodyfikowanego łańcucha
+        //Teraz definiujemy test wykrywania manipulacji zawartości bloku
         [TestMethod]
-        public void CzyWykrywaManipulacjeZawartosci()
+        public void CzyWykrywaZmianeZawartosci()
         {
             //Arrange
             blockchain.AddNewVersion("A");
@@ -109,30 +109,40 @@ namespace BlockchainVersionControl.Tests
             Assert.IsFalse(blockchain.IsChainValid());
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy SaveToFile i LoadFromFile działają poprawnie
+        //Definiujemy test poprawności działania zapisu i odczytu z pliku
         [TestMethod]
-        public void CzyZapisujeIOdczytujePoprawnieLancuch()
+        public void CzyZapisujeIOdczytujePoprawnie()
         {
-            //Arrange
             var sciezka = "test_chain.json";
-            blockchain.AddNewVersion("Linia nowa");
-            blockchain.SaveToFile(sciezka);
+            try
+            {
+                //Arrange
+                blockchain.AddNewVersion("Linia nowa");
+                blockchain.SaveToFile(sciezka);
 
-            //Act
-            var wczytany = VersionControlBlockchain.LoadFromFile(sciezka);
+                //Act
+                var wczytany = VersionControlBlockchain.LoadFromFile(sciezka);
 
-            //Assert
-            Assert.AreEqual(blockchain.Chain.Count, wczytany.Chain.Count);
-            File.Delete(sciezka);
+                //Assert
+                Assert.AreEqual(blockchain.Chain.Count, wczytany.Chain.Count);
+                Assert.AreEqual(blockchain.Chain[1].Hash, wczytany.Chain[1].Hash);
+            }
+            finally
+            {
+                if (File.Exists(sciezka))
+                    File.Delete(sciezka);
+            }
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy LoadFromFile obsługuje pusty plik bez wyjątku
+        //Teraz definiujemy test odczytu pustego pliku JSON
         [TestMethod]
-        public void CzyWczytujePustyLubBlednyPlik()
+        public void CzyWczytujePustyPlik()
         {
-            //Arrange i Act
+            //Arrange
             var sciezka = "empty.json";
             File.WriteAllText(sciezka, "");
+
+            //Act
             var wczytany = VersionControlBlockchain.LoadFromFile(sciezka);
 
             //Assert
@@ -140,9 +150,9 @@ namespace BlockchainVersionControl.Tests
             File.Delete(sciezka);
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy IsChainValid wykrywa niepoprawny hash poprzedniego bloku
+        //Przechodzimy do zdefiniowania testu wykrywania błędnego poprzedniego hasha w łańcuchu bloków
         [TestMethod]
-        public void CzyWykrywaNiepoprawnyHashPoprzedniegoBloku()
+        public void CzyWykrywaBlednyPoprzedniHash()
         {
             //Arrange
             blockchain.AddNewVersion("Nowa linia");
@@ -156,40 +166,63 @@ namespace BlockchainVersionControl.Tests
             Assert.IsFalse(czyPoprawny);
         }
 
-        //Definiujemy test, który odpowiada za sprawdzenie czy CalculateHash oblicza poprawny hash
+        //Definiujemy test porównania hashy przy zmianie zawartości bloku
         [TestMethod]
-        public void IsChainValid_ReturnsFalse_WhenPreviousHashDoesNotMatch()
+        public void CzyZmienionyDiffZmieniaHash()
         {
-            // Arrange
-            var blockchain = new VersionControlBlockchain("Initial content");
-            blockchain.AddNewVersion("First update");
-            blockchain.AddNewVersion("Second update");
-
-            // Manipulacja: zmieniamy PreviousHash drugiego bloku
-            blockchain.Chain[2].PreviousHash = "tampered_hash";
-
-            // Act
-            bool isValid = blockchain.IsChainValid();
-
-            // Assert
-            Assert.IsFalse(isValid, "Łańcuch powinien być nieważny, gdy PreviousHash nie odpowiada Hash poprzedniego bloku.");
-        }
-
-        //Definiujemy test, który odpowiada za sprawdzenie czy CalculateHash oblicza poprawny hash
-        [TestMethod]
-        public void CalculateHash_ReturnsDifferentHash_WhenContentChanges()
-        {
-            // Arrange
+            //Arrange
             var timestamp = DateTime.Now;
             var block1 = new DocumentVersion(1, timestamp, "Diff1", "prevHash");
             var block2 = new DocumentVersion(1, timestamp, "Diff2", "prevHash");
 
-            // Act
-            var hash1 = block1.Hash;
-            var hash2 = block2.Hash;
+            //Assert
+            Assert.AreNotEqual(block1.Hash, block2.Hash);
+        }
 
-            // Assert
-            Assert.AreNotEqual(hash1, hash2, "Zmiana zawartości bloku powinna prowadzić do zmiany jego hasha.");
+        //Definiujemy test porównania hashy przy zmianie czasu utworzenia bloku
+        [TestMethod]
+        public void CzyZmienionyCzasZmieniaHash()
+        {
+            //Arrange
+            var block1 = new DocumentVersion(1, DateTime.UtcNow, "Zawartosc", "poprzedniHash");
+            System.Threading.Thread.Sleep(10);
+
+            //Act
+            var block2 = new DocumentVersion(1, DateTime.UtcNow, "Zawartosc", "poprzedniHash");
+
+            //Assert
+            Assert.AreNotEqual(block1.Hash, block2.Hash);
+        }
+
+        //Teraz definiujemy test zwracania wyjątku przy żądaniu wersji o nieistniejącym indeksie
+        [TestMethod]
+        public void CzyZglaszaWyjatekDlaZlegoIndeksu()
+        {
+            //Assert
+            Assert.ThrowsException<ArgumentOutOfRangeException>(() => blockchain.GetDocumentVersion(999));
+        }
+
+        //Definiujemy test poprawności łańcucha przy braku danych w Diff
+        [TestMethod]
+        public void CzyWykrywaNullDiff()
+        {
+            //Arrange
+            blockchain.AddNewVersion("Tekst");
+            blockchain.Chain[1].Diff = null;
+
+            //Assert
+            Assert.IsFalse(blockchain.IsChainValid());
+        }
+
+        //Definiujemy test poprawności łańcucha przy braku danych w PreviousHash
+        [TestMethod]
+        public void CzyZwracaOstatniBlokNaPoczatku()
+        {
+            //Arrange i Act
+            var ostatni = blockchain.GetLatestVersionBlock();
+
+            //Assert
+            Assert.AreEqual(0, ostatni.Index);
         }
 
     }
